moved roll command into its own class
