+++ src/main/java/dansplugins/rpsystem/CommandInterpreter.java
@@ -106,24 +106,8 @@ public class CommandInterpreter {
         }
 
         if (label.equalsIgnoreCase("roll") || label.equalsIgnoreCase("dice")) {
-            if (sender instanceof Player) {
-                Player player = (Player) sender;
-                if (player.hasPermission("rp.roll") || player.hasPermission("rp.dice") || player.hasPermission("rp.default")) {
-                    if (args.length > 0) {
-                        try {
-                            int max = Integer.parseInt(args[0]);
-                            Messenger.getInstance().sendMessageToPlayersWithinDistance(player,ChatColor.AQUA + "" + ChatColor.ITALIC + player.getName() + " has rolled a " + rollDice(max) + " out of " + max + ".", 25);
-                        }
-                        catch(Exception ignored) {
-
-                        }
-                    }
-                }
-                else {
-                    player.sendMessage(ChatColor.RED + "Sorry! In order to use this command, you need one the following permissions: 'rp.roll', 'rp.dice'");
-                }
-
-            }
+            RollCommand command = new RollCommand();
+            return command.rollDice(sender, args);
         }
 
         if (label.equalsIgnoreCase("title")) {
@@ -151,10 +135,6 @@ public class CommandInterpreter {
         }
 
         return false;
-    }
-
-    private static int rollDice(int max) {
-        return (int)(Math.random() * max + 1);
     }
 
 }
