+++ src/test/java/dansplugins/rpsystem/commands/RollCommandTest.java
@@ -6,7 +6,7 @@ import org.mockito.Mock;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
-import static dansplugins.rpsystem.commands.RollCommand.*;
+import static dansplugins.rpsystem.commands.NewRollCommand.*;
 
 public class RollCommandTest {
